echomimic_v2/requirements.txt
transformers>=4.46.3
diffusers==0.31.0
torchmetrics
torchtyping
tqdm
einops==0.8.0
omegaconf==2.3.0
opencv-python
av==13.1.0
gradio
accelerate==1.1.1
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
decord==0.6.0
gradio_client==1.4.3
imageio==2.36.0
imageio-ffmpeg==0.5.1
numpy==1.26.4
onnxruntime-gpu==1.20.1
open-clip-torch==2.29.0
opencv-contrib-python==4.10.0.84
Pillow<10.3.0,>=10.2.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
torchdiffeq==0.2.5
torchsde==0.2.6
mlflow==2.18.0
controlnet-aux==0.0.9
ffmpeg-python
soundfile
mediapipe
IPython
scenedetect
moviepy==1.0.3
huggingface_hub==0.26.2
mediapipe
